embedding/requirements.txt
chromadb==0.4.23
GitPython==3.1.42
langchain-community==0.0.24
langchain-openai==0.0.7
langchain==0.1.9
mypy==1.8.0
pylint==3.0.4
python-dotenv==1.0.1
tiktoken==0.6.0
tokenizers==0.15.2
tree-sitter-languages==1.10.2
tree-sitter==0.20.4
esprima==4.0.1
